Feature: Multi-set convergence - maybe!
When checking the number of active neurons against the desired number of active neurons in the run() method, the conditional should be an equivalency statement (I believe). But currently this breaks convergence when given multiple categories of neurons (not sure about single category networks).

TODO: Check that single-category networks still works with the newly added overloaded constructor.